refactor: rename factories folder to singular
