+++ src/etl/factory/ResearchStudyModelFactory.test.ts
