+++ src/etl/factory/ResearchStudyModelFactory.ts
