+++ src/etl/ingest-pipelines/IngestPipelineDmDmdiv.ts
@@ -2,7 +2,7 @@ import { IngestPipeline, IndexElasticsearch, ResearchStudyElasticsearchDocument 
 import { ResearchStudyModel } from '../../shared/models/domain-resources/ResearchStudyModel'
 import { EclaireDto } from '../dto/EclaireDto'
 import { RiphDmDto } from '../dto/RiphDmDto'
-import { ResearchStudyModelFactory } from '../factories/ResearchStudyModelFactory'
+import { ResearchStudyModelFactory } from '../factory/ResearchStudyModelFactory'
 
 export class IngestPipelineDmDmdiv extends IngestPipeline {
   readonly type = 'dm-dmdiv'
