+++ src/etl/ingest-pipelines/IngestPipelineJarde.ts
@@ -2,7 +2,7 @@ import { IngestPipeline, IndexElasticsearch, ResearchStudyElasticsearchDocument 
 import { ResearchStudyModel } from '../../shared/models/domain-resources/ResearchStudyModel'
 import { EclaireDto } from '../dto/EclaireDto'
 import { RiphJardeDto } from '../dto/RiphJardeDto'
-import { ResearchStudyModelFactory } from '../factories/ResearchStudyModelFactory'
+import { ResearchStudyModelFactory } from '../factory/ResearchStudyModelFactory'
 
 export class IngestPipelineJarde extends IngestPipeline {
   readonly type = 'jarde'
